# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateTaskRequest:
    properties:
      description:
        minLength: 1
        type: string
      difficulty:
        allOf:
        - $ref: '#/definitions/models.Difficulty'
        enum:
        - easy
        - medium
        - hard
      initial_code:
        type: string
      test_cases:
        items:
          $ref: '#/definitions/models.TestCase'
        minItems: 1
        type: array
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - description
    - difficulty
    - test_cases
    - title
    type: object
  models.Difficulty:
    enum:
    - easy
    - medium
    - hard
    type: string
    x-enum-varnames:
    - DifficultyEasy
    - DifficultyMedium
    - DifficultyHard
  models.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/models.ErrorBody'
    type: object
...
  models.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  models.Role:
    enum:
    - student
    - admin
    type: string
    x-enum-varnames:
    - RoleStudent
    - RoleAdmin
...
  models.Task:
    properties:
      created_at:
        type: string
      description:
        type: string
      difficulty:
        $ref: '#/definitions/models.Difficulty'
      id:
        type: integer
      initial_code:
        type: string
      lesson_id:
        type: integer
      test_cases:
        items:
          type: object
        type: array
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.TestCase:
    properties:
      expected:
        type: string
      input:
        type: string
    type: object
...
  models.UpdateTaskRequest:
    properties:
      description:
        minLength: 1
        type: string
      difficulty:
        allOf:
        - $ref: '#/definitions/models.Difficulty'
        enum:
        - easy
        - medium
        - hard
      initial_code:
        type: string
      test_cases:
        items:
          $ref: '#/definitions/models.TestCase'
        minItems: 1
        type: array
      title:
        maxLength: 200
        minLength: 1
        type: string
    type: object
  models.UpdateUserRequest:
    properties:
      name:
        maxLength: 100
        minLength: 1
        type: string
    type: object
  models.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        $ref: '#/definitions/models.Role'
    type: object
host: localhost:8080
info:
  contact: {}
  description: Online programming learning platform API
  title: Diplom API
  version: "1.0"
paths:
...
  /lessons/{lesson_id}/tasks:
    get:
      description: Get a paginated list of tasks for a lesson
      parameters:
      - description: Lesson ID
        in: path
        name: lesson_id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List tasks in a lesson
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      description: Create a new task in a lesson
      parameters:
      - description: Lesson ID
        in: path
        name: lesson_id
        required: true
        type: integer
      - description: Task data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a task (admin)
      tags:
      - Tasks
...
  /tasks/{id}:
    delete:
      description: Delete task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a task (admin)
      tags:
      - Tasks
    get:
      description: Get task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a task
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      description: Update task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a task (admin)
      tags:
      - Tasks
...
  /users:
    get:
      description: Get a paginated list of all users
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search by name or email
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users (admin)
      tags:
      - Users
  /users/{id}:
    delete:
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user (admin)
      tags:
      - Users
  /users/me:
    get:
      description: Get the profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update the profile of the authenticated user
      parameters:
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update current user
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"